o [engine] Added EFFECT_RESTORE to support mana restoration potions. Had to   sacrifice EFFECT_NATURAL because the effects bitmask is out of bits. Turns   out bit 31 does not work right because of an issue with atoi interpreting it   as an overflow. This will be addressed in the 0.3.x line with the new effects   system.

+++ src/game.c
@@ -75,7 +75,7 @@ static struct terrain_palette *LTP_wrapper(class Loader * loader);
     ret = -1; \
     goto cleanup; \
    } \
-  (var) = atoi(lexer_lexeme(Lexer));
+  (var) = atol(lexer_lexeme(Lexer));
 
 #define PARSE_CHAR(tag,var)         \
   if (! MATCH_WORD((tag))) { \
